+++ familyhelper-webapi-node/src/main/java/com/dwarfeng/familyhelper/webapi/node/controller/v1/life/PoatController.java
@@ -1,0 +1,143 @@
+package com.dwarfeng.familyhelper.webapi.node.controller.v1.life;
+
+import com.dwarfeng.familyhelper.life.sdk.bean.entity.JSFixedFastJsonPoat;
+import com.dwarfeng.familyhelper.life.stack.bean.entity.Poat;
+import com.dwarfeng.familyhelper.life.stack.bean.key.PoatKey;
+import com.dwarfeng.familyhelper.webapi.sdk.bean.disp.life.JSFixedFastJsonDispPoat;
+import com.dwarfeng.familyhelper.webapi.stack.bean.disp.life.DispPoat;
+import com.dwarfeng.familyhelper.webapi.stack.handler.system.TokenHandler;
+import com.dwarfeng.familyhelper.webapi.stack.service.life.PoatResponseService;
+import com.dwarfeng.subgrade.sdk.bean.dto.FastJsonResponseData;
+import com.dwarfeng.subgrade.sdk.bean.dto.JSFixedFastJsonPagedData;
+import com.dwarfeng.subgrade.sdk.bean.dto.PagingUtil;
+import com.dwarfeng.subgrade.sdk.bean.dto.ResponseDataUtil;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.SkipRecord;
+import com.dwarfeng.subgrade.sdk.interceptor.login.LoginRequired;
+import com.dwarfeng.subgrade.stack.bean.BeanTransformer;
+import com.dwarfeng.subgrade.stack.bean.dto.PagedData;
+import com.dwarfeng.subgrade.stack.bean.dto.PagingInfo;
+import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
+import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
+import com.dwarfeng.subgrade.stack.exception.ServiceExceptionMapper;
+import org.springframework.web.bind.annotation.*;
+
+import javax.servlet.http.HttpServletRequest;
+
+/**
+ * 活动模板权限响应控制器。
+ *
+ * @author DwArFeng
+ * @since 1.0.10
+ */
+@RestController
+@RequestMapping("/api/v1/life")
+public class PoatController {
+
+    private final PoatResponseService service;
+
+    private final ServiceExceptionMapper sem;
+
+    private final BeanTransformer<Poat, JSFixedFastJsonPoat> poatBeanTransformer;
+    private final BeanTransformer<DispPoat, JSFixedFastJsonDispPoat> dispPoatBeanTransformer;
+
+    private final TokenHandler tokenHandler;
+
+    public PoatController(
+            PoatResponseService service,
+            ServiceExceptionMapper sem,
+            BeanTransformer<Poat, JSFixedFastJsonPoat> poatBeanTransformer,
+            BeanTransformer<DispPoat, JSFixedFastJsonDispPoat> dispPoatBeanTransformer,
+            TokenHandler tokenHandler
+    ) {
+        this.service = service;
+        this.sem = sem;
+        this.poatBeanTransformer = poatBeanTransformer;
+        this.dispPoatBeanTransformer = dispPoatBeanTransformer;
+        this.tokenHandler = tokenHandler;
+    }
+
+    @GetMapping("/poat/{longId}&{stringId}/exists")
+    @BehaviorAnalyse
+    @LoginRequired
+    public FastJsonResponseData<Boolean> exists(
+            HttpServletRequest request, @PathVariable("longId") Long longId, @PathVariable("stringId") String stringId
+    ) {
+        try {
+            boolean exists = service.exists(new PoatKey(longId, stringId));
+            return FastJsonResponseData.of(ResponseDataUtil.good(exists));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(e, sem));
+        }
+    }
+
+    @GetMapping("/poat/{longId}&{stringId}")
+    @BehaviorAnalyse
+    @LoginRequired
+    public FastJsonResponseData<JSFixedFastJsonPoat> get(
+            HttpServletRequest request, @PathVariable("longId") Long longId, @PathVariable("stringId") String stringId
+    ) {
+        try {
+            Poat poat = service.get(new PoatKey(longId, stringId));
+            return FastJsonResponseData.of(ResponseDataUtil.good(JSFixedFastJsonPoat.of(poat)));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(e, sem));
+        }
+    }
+
+    @GetMapping("/poat/{longId}&{stringId}/disp")
+    @BehaviorAnalyse
+    @LoginRequired
+    public FastJsonResponseData<JSFixedFastJsonDispPoat> getDisp(
+            HttpServletRequest request, @PathVariable("longId") Long longId, @PathVariable("stringId") String stringId
+    ) {
+        try {
+            StringIdKey inspectAccountKey = tokenHandler.getAccountKey(request);
+            DispPoat dispPoat = service.getDisp(new PoatKey(longId, stringId), inspectAccountKey);
+            return FastJsonResponseData.of(ResponseDataUtil.good(JSFixedFastJsonDispPoat.of(dispPoat)));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(e, sem));
+        }
+    }
+
+    @GetMapping("/activity-template/{activityTemplateId}/poat")
+    @BehaviorAnalyse
+    @SkipRecord
+    @LoginRequired
+    public FastJsonResponseData<JSFixedFastJsonPagedData<JSFixedFastJsonPoat>> childForActivityTemplate(
+            HttpServletRequest request, @PathVariable("activityTemplateId") Long activityTemplateId,
+            @RequestParam("page") int page, @RequestParam("rows") int rows) {
+        try {
+            PagedData<Poat> childForActivityTemplate = service.childForActivityTemplate(
+                    new LongIdKey(activityTemplateId), new PagingInfo(page, rows)
+            );
+            PagedData<JSFixedFastJsonPoat> transform = PagingUtil.transform(
+                    childForActivityTemplate, poatBeanTransformer
+            );
+            return FastJsonResponseData.of(ResponseDataUtil.good(JSFixedFastJsonPagedData.of(transform)));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(e, sem));
+        }
+    }
+
+    @GetMapping("/activity-template/{activityTemplateId}/poat/disp")
+    @BehaviorAnalyse
+    @SkipRecord
+    @LoginRequired
+    public FastJsonResponseData<JSFixedFastJsonPagedData<JSFixedFastJsonDispPoat>> childForActivityTemplateDisp(
+            HttpServletRequest request, @PathVariable("activityTemplateId") Long activityTemplateId,
+            @RequestParam("page") int page, @RequestParam("rows") int rows) {
+        try {
+            StringIdKey inspectAccountKey = tokenHandler.getAccountKey(request);
+            PagedData<DispPoat> childForActivityTemplateDisp = service.childForActivityTemplateDisp(
+                    new LongIdKey(activityTemplateId), new PagingInfo(page, rows), inspectAccountKey
+            );
+            PagedData<JSFixedFastJsonDispPoat> transform = PagingUtil.transform(
+                    childForActivityTemplateDisp, dispPoatBeanTransformer
+            );
+            return FastJsonResponseData.of(ResponseDataUtil.good(JSFixedFastJsonPagedData.of(transform)));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(e, sem));
+        }
+    }
+}