+++ CHANGELOG.md
@@ -22,6 +22,7 @@
   - com.dwarfeng.familyhelper.webapi.node.controller.v1.life.ActivityDataNodeController。
   - com.dwarfeng.familyhelper.webapi.node.controller.v1.life.ActivityDataRecordController。
   - com.dwarfeng.familyhelper.webapi.node.controller.v1.life.ActivityTemplateController。
+  - com.dwarfeng.familyhelper.webapi.node.controller.v1.life.PoatController。
 
 - 为 dubbo 的消费者增加了分组设置。
   - snowflake。